requirements.txt
requests~=2.32.5
yfinance~=0.2.65
protobuf~=6.32.0
google-api-python-client~=2.183.0
python-dotenv~=1.2.1